test(frontend): consolidate SGLang skip-special-token coverage
Fold the decode-option regression tests into test_sglang_processor_unit.py
and remove the standalone test module, as requested in PR #14602 review.

Parameterize one generator test over omitted, true, and false request
values and inline/pool execution. Assert both returned text and the
routed output option. Exercise the real preprocessing worker through a
thread pool without loading a model or spawning subprocesses.

Reuse the existing parser tests for delimiter preservation. Set
skip_special_tokens=False in the existing request stop-token test so
matched-stop trimming remains covered when token display is enabled,
and remove the redundant stop-trimming parametrization.
